+++ src/library/DIPS.Mobile.UI/Components/Searching/SearchBar.Properties.cs
@@ -292,8 +292,13 @@ namespace DIPS.Mobile.UI.Components.Searching
         public static readonly BindableProperty SearchFieldBackgroundColorProperty = BindableProperty.Create(
             nameof(SearchFieldBackgroundColor),
             typeof(Color),
-            typeof(SearchBar),
-            defaultValue: DIPS.Mobile.UI.Resources.Colors.Colors.GetColor(ColorName.color_fill_neutral));      
+            typeof(SearchBar), 
+#if __IOS__ 
+            // iOS places a semi-transparent background on top of the search field by default, so we use a different default value for iOS
+            defaultValue: DIPS.Mobile.UI.Resources.Colors.Colors.GetColor(ColorName.color_palette_neutral_white));    
+#else 
+            defaultValue: DIPS.Mobile.UI.Resources.Colors.Colors.GetColor(ColorName.color_fill_neutral));
+#endif
         
         public static readonly BindableProperty PlaceholderColorProperty = BindableProperty.Create(
             nameof(PlaceholderColor),